refactor: move carbon estimator form to components
