+++ src/app/components/carbon-estimator-form/carbon-estimator-form.constants.ts
@@ -1,6 +1,6 @@
 import { FormGroup } from '@angular/forms';
-import { ExpansionPanelConfig } from '../components/expansion-panel/expansion-panel.constants';
-import { CostRange, EstimatorFormValues, WorldLocation } from '../types/carbon-estimator';
+import { ExpansionPanelConfig } from '../expansion-panel/expansion-panel.constants';
+import { CostRange, EstimatorFormValues, WorldLocation } from '../../types/carbon-estimator';
 
 export const costRanges: CostRange[] = [
   { min: 0, max: 1000 },
