+++ src/app/components/error-summary/error-summary.component.ts
@@ -1,5 +1,5 @@
 import { Component, ElementRef, input, ViewChild } from '@angular/core';
-import { ValidationError } from '../../carbon-estimator-form/carbon-estimator-form.constants';
+import { ValidationError } from '../../components/carbon-estimator-form/carbon-estimator-form.constants';
 
 @Component({
   selector: 'error-summary',
