+++ src/app/components/location-input/location-input.component.ts
@@ -4,7 +4,7 @@ import {
   Location,
   locationDescriptions,
   questionPanelConfig,
-} from '../../carbon-estimator-form/carbon-estimator-form.constants';
+} from '../../components/carbon-estimator-form/carbon-estimator-form.constants';
 import { ExpansionPanelComponent } from '../expansion-panel/expansion-panel.component';
 import { locationArray, WorldLocation } from '../../types/carbon-estimator';
 import { ControlValueAccessor, FormControl, NG_VALUE_ACCESSOR, ReactiveFormsModule } from '@angular/forms';
