+++ src/app/features/customers/components/customer-form-section.component.ts
@@ -4,7 +4,7 @@ import {
   formContext,
   FormContextSection,
   questionPanelConfig,
-} from '../../../carbon-estimator-form/carbon-estimator-form.constants';
+} from '../../../components/carbon-estimator-form/carbon-estimator-form.constants';
 import { InvalidatedPipe } from '../../../pipes/invalidated.pipe';
 import { LocationInputComponent } from '../../../components/location-input/location-input.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
