+++ src/app/features/organisation/components/organisation-form-section.component.ts
@@ -1,6 +1,9 @@
 import { Component, computed, inject, OnInit, Signal, signal } from '@angular/core';
 import { SectionHeaderComponent } from '../../../components/section-header/section-header.component';
-import { formContext, FormContextSection } from '../../../carbon-estimator-form/carbon-estimator-form.constants';
+import {
+  formContext,
+  FormContextSection,
+} from '../../../components/carbon-estimator-form/carbon-estimator-form.constants';
 import { defaultValues, OrganisationFormGroup, OrganisationFormService } from '../services/organisation-form.service';
 import { InvalidatedPipe } from '../../../pipes/invalidated.pipe';
 import { LocationInputComponent } from '../../../components/location-input/location-input.component';
