+++ src/app/services/form.service.ts
@@ -2,7 +2,7 @@ import { inject, Injectable } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 import { defaultSaasValues, SaasFormService } from '../features/saas/services/saas-form.service';
 import { EstimatorFormValues, EstimatorValues } from '../types/carbon-estimator';
-import { EstimatorFormRawValue } from '../carbon-estimator-form/carbon-estimator-form.constants';
+import { EstimatorFormRawValue } from '../components/carbon-estimator-form/carbon-estimator-form.constants';
 import { CloudFormService, defaultCloudValues } from '../features/cloud/services/cloud-form.service';
 import { OrganisationFormService } from '../features/organisation/services/organisation-form.service';
 import { defaultOnPremValues, OnPremiseFormService } from '../features/on-premise/services/on-premise-form.service';
